move scaling to own class
